replace stdatomic with WINAPI Interlocked* functions
the main thread only reads 'keepRunning' once every 'thread_tick_ms'
(100ms), so a race between the main thread and the thread that handles
the SIGINT (and where 'keepRunning' is updated) may never happen,
but still it can, in theory.

the previous stdatomic solution was pretty much a left over attempt
and did nothing. also stdatomic is more of an extension.

of course, the correct way of avoiding a race in a WINAPI way is to
use the Interlocked* functions. this patch adds just that.