ping_timeout: track the current state with an enum variable
The error / success states are now tracked via the "enum STATE_SWITCH"
typed variable "state_switch". The state is not know in the begining,
so the value of "state_switch" is ST_NONE.

This change makes sure that an error sound is played if the
ICMP echo fails on the first call.